Set default WSL version to 1 before install (#28)
In windows-2022 environment WSLv2 is now generally available and the default,
but it cannot work as GitHub actions are running on Azure cloud where nested
Hyper-V virtualization is not available.

GitHub will change the default WSL version back to WSLv1 in a future update
but to be future proof and not break again in case WSLv2 becomes properly
available in the future, we force it additionally in the action.