Remove usage of Dukat
Dukat results are still very suboptimal and in Kotlin 1.8.20-Beta it was disabled.
As most usages were already replaced by kotlin-wrappers or self-written class,
now the last remaining three classes are hand-written too and Dukat is not necessary,
including the massive post-processing just to make it compile.
This will save the time for running Dukat, for doing the post processing,
and for compiling quite some unncessary classes.
If Dukat is improved and added back to the Gradle plugin integration,
this decision can be reconsidered, but for now it saves much time and is better
future proof in the light of 1.8.20-Beta.